Separate start scripts from models (#33)
