+++ .vscode/launch.json
@@ -2,14 +2,14 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "start",
+            "name": "register",
             "type": "node",
             "request": "launch",
             "protocol": "inspector",
             "preLaunchTask": "build",
             "cwd": "${workspaceFolder}",
             "runtimeExecutable": "node",
-            "args": ["--enable-source-maps", "${workspaceFolder}/dist/start.js"],
+            "args": ["--enable-source-maps", "${workspaceFolder}/dist/start-bot.js", "--register"],
             "resolveSourceMapLocations": ["${workspaceFolder}/**", "!**/node_modules/**"],
             "outputCapture": "std",
             "internalConsoleOptions": "openOnSessionStart",
@@ -17,14 +17,29 @@
             "restart": false
         },
         {
-            "name": "start:shard",
+            "name": "start:bot",
             "type": "node",
             "request": "launch",
             "protocol": "inspector",
             "preLaunchTask": "build",
             "cwd": "${workspaceFolder}",
             "runtimeExecutable": "node",
-            "args": ["--enable-source-maps", "${workspaceFolder}/dist/app.js"],
+            "args": ["--enable-source-maps", "${workspaceFolder}/dist/start-bot.js"],
+            "resolveSourceMapLocations": ["${workspaceFolder}/**", "!**/node_modules/**"],
+            "outputCapture": "std",
+            "internalConsoleOptions": "openOnSessionStart",
+            "skipFiles": ["<node_internals>/**"],
+            "restart": false
+        },
+        {
+            "name": "start:manager",
+            "type": "node",
+            "request": "launch",
+            "protocol": "inspector",
+            "preLaunchTask": "build",
+            "cwd": "${workspaceFolder}",
+            "runtimeExecutable": "node",
+            "args": ["--enable-source-maps", "${workspaceFolder}/dist/start-manager.js"],
             "resolveSourceMapLocations": ["${workspaceFolder}/**", "!**/node_modules/**"],
             "outputCapture": "std",
             "internalConsoleOptions": "openOnSessionStart",
