+++ package.json
@@ -10,8 +10,8 @@
     },
     "type": "module",
     "exports": [
-        "./dist/start.js",
-        "./dist/app.js"
+        "./dist/start-bot.js",
+        "./dist/start-manager.js"
     ],
     "scripts": {
         "lint": "eslint . --cache --ext .js,.jsx,.ts,.tsx",
@@ -21,10 +21,11 @@
         "clean": "git clean -xdf --exclude=\"/config/**/*\"",
         "clean:dry": "git clean -xdf --exclude=\"/config/**/*\" --dry-run",
         "build": "tsc --project tsconfig.json",
-        "register": "npm run build && node --enable-source-maps dist/start.js --register",
-        "register:clear": "npm run build && node --enable-source-maps dist/start.js --register:clear",
-        "start": "npm run build && node --enable-source-maps dist/start.js",
-        "start:shard": "npm run build && node --enable-source-maps dist/app.js",
+        "register": "npm run build && node --enable-source-maps dist/start-bot.js --register",
+        "register:clear": "npm run build && node --enable-source-maps dist/start-bot.js --register:clear",
+        "start": "npm run start:bot",
+        "start:bot": "npm run build && node --enable-source-maps dist/start-bot.js",
+        "start:manager": "npm run build && node --enable-source-maps dist/start-manager.js",
         "start:pm2": "npm run build && npm run pm2:start",
         "pm2:start": "pm2 start process.json",
         "pm2:stop": "pm2 stop process.json",
