+++ process.json
@@ -2,7 +2,7 @@
     "apps": [
         {
             "name": "my-bot",
-            "script": "dist/app.js",
+            "script": "dist/start-manager.js",
             "node_args": ["--enable-source-maps"],
             "restart_delay": 10000
         }
