+++ src/models/api.ts
@@ -2,13 +2,13 @@ import express, { Express } from 'express';
 import { createRequire } from 'node:module';
 import util from 'node:util';
 
-import { Controller } from './controllers/index.js';
-import { checkAuth, handleError } from './middleware/index.js';
-import { Logger } from './services/index.js';
+import { Controller } from '../controllers/index.js';
+import { checkAuth, handleError } from '../middleware/index.js';
+import { Logger } from '../services/index.js';
 
 const require = createRequire(import.meta.url);
-let Config = require('../config/config.json');
-let Logs = require('../lang/logs.json');
+let Config = require('../../config/config.json');
+let Logs = require('../../lang/logs.json');
 
 export class Api {
     private app: Express;
