+++ src/models/manager.ts
@@ -1,12 +1,12 @@
 import { Shard, ShardingManager } from 'discord.js';
 import { createRequire } from 'node:module';
 
-import { JobService, Logger } from './services/index.js';
+import { JobService, Logger } from '../services/index.js';
 
 const require = createRequire(import.meta.url);
-let Config = require('../config/config.json');
-let Debug = require('../config/debug.json');
-let Logs = require('../lang/logs.json');
+let Config = require('../../config/config.json');
+let Debug = require('../../config/debug.json');
+let Logs = require('../../lang/logs.json');
 
 export class Manager {
     constructor(private shardManager: ShardingManager, private jobService: JobService) {}
