+++ src/start-manager.ts
@@ -2,10 +2,10 @@ import { ShardingManager } from 'discord.js';
 import { createRequire } from 'node:module';
 
 import 'reflect-metadata';
-import { Api } from './api.js';
 import { GuildsController, RootController, ShardsController } from './controllers/index.js';
 import { Job, UpdateServerCountJob } from './jobs/index.js';
-import { Manager } from './manager.js';
+import { Api } from './models/api.js';
+import { Manager } from './models/manager.js';
 import { HttpService, JobService, Logger, MasterApiService } from './services/index.js';
 import { MathUtils, ShardUtils } from './utils/index.js';
 
@@ -51,7 +51,7 @@ async function start(): Promise<void> {
         return;
     }
 
-    let shardManager = new ShardingManager('dist/start.js', {
+    let shardManager = new ShardingManager('dist/start-bot.js', {
         token: Config.client.token,
         mode: Debug.override.shardMode.enabled ? Debug.override.shardMode.value : 'worker',
         respawn: true,
